Tidy up comments
